feat(tests): Add display utility functions and macros for test results
